player partialize setup
