+++ Assets/Code/Fsm/GravityFsm/PlayerFsm/PlayerFsm_Dashsquat.cs
@@ -5,4 +5,17 @@ public partial class PlayerFsm
         HandleCollisionMove();
         HandleTurning(DashsquatTurnMultiplier, true);
     }
+
+    private void DashsquatConfigure()
+    {
+        Machine.Configure(PlayerFsmState.Dashsquat)
+            .SubstateOf(GravityFsmState.Aerial)
+            .Permit(FsmTrigger.Timeout, PlayerFsmState.Grapple)
+            .SubstateOf(GravityFsmState.DontApplyYVelocity)
+            .OnEntry(_ =>
+            {
+                _inputBuffer.ConsumeBuffer("Dash");
+                ReplaceAnimatorTrigger("Dashsquat");
+            });
+    }
 }
