+++ Assets/Code/Fsm/GravityFsm/PlayerFsm/PlayerFsm_Fall.cs
@@ -1,0 +1,23 @@
+public partial class PlayerFsm
+{
+    private void FallConfigure()
+    {
+        Machine.Configure(PlayerFsmState.Fall)
+            .SubstateOf(GravityFsmState.Aerial)
+            .Permit(GravityFsmTrigger.StartFrameGrounded, PlayerFsmState.Landsquat)
+            .PermitIf(PlayerFsmTrigger.Jump, PlayerFsmState.Jumpsquat, _ => TimeInAir <= CoyoteTime)
+            .PermitIf(GravityFsmTrigger.StartFrameGrounded, PlayerFsmState.HardLand, _ => AirYDiff() < HardLandAirDiff,
+                1)
+            .PermitIf(GravityFsmTrigger.StartFrameGrounded, PlayerFsmState.HardLandRoll,
+                _ => AirYDiff() < HardLandAirDiff && _momentum > HardLandRollMinimumMomentum, 2)
+            .Permit(PlayerFsmTrigger.Dash, PlayerFsmState.Dashsquat)
+            .OnEntry(_ => { ReplaceAnimatorTrigger("Fall"); });
+
+        Machine.Configure(PlayerFsmState.HardTurn)
+            .SubstateOf(PlayerFsmState.LockMomentum)
+            .Permit(PlayerFsmTrigger.NoMomentum, PlayerFsmState.GroundMove)
+            .Permit(GravityFsmTrigger.StartFrameAerial, PlayerFsmState.Fall)
+            .SubstateOf(GravityFsmState.Grounded)
+            .OnEntry(_ => { ReplaceAnimatorTrigger("HardTurn"); });
+    }
+}
