+++ Assets/Code/Fsm/GravityFsm/PlayerFsm/PlayerFsm_HardLandRoll.cs
@@ -6,4 +6,16 @@ public partial class PlayerFsm
     {
         transform.position += ComputeCollisionMove(transform.forward * (HardLandRollForwardSpeed * Time.deltaTime));
     }
+
+    private void HardLandRollConfigure()
+    {
+        Machine.Configure(PlayerFsmState.HardLandRoll)
+            .SubstateOf(GravityFsmState.Grounded)
+            .Permit(FsmTrigger.Timeout, PlayerFsmState.GroundMove)
+            .OnEntry(_ =>
+            {
+                _momentum = HardLandRollExitMomentum;
+                ReplaceAnimatorTrigger("HardLandRoll");
+            });
+    }
 }
