+++ Assets/Code/Fsm/GravityFsm/PlayerFsm/PlayerFsm_ImpaleGround.cs
@@ -20,4 +20,22 @@ public partial class PlayerFsm
             : Mathf.Max(_stateEntryMomentum + ImpaleMomentumOffset, ImpaleMinimumMomentumAfterOffset);
         _momentum = Mathf.Lerp(_momentum, targetMomentum, Time.deltaTime * ImpaleMomentumLerpStrenth);
     }
+
+    private void ImpaleGroundConfigure()
+    {
+        Machine.Configure(PlayerFsmState.ImpaleGround)
+            .SubstateOf(GravityFsmState.Grounded)
+            .Permit(FsmTrigger.Timeout, PlayerFsmState.GroundMove)
+            .Permit(GravityFsmTrigger.StartFrameAerial, PlayerFsmState.Fall)
+            .Permit(PlayerFsmTrigger.Jump, PlayerFsmState.Jumpsquat)
+            .PermitIf(PlayerFsmTrigger.Attack, PlayerFsmState.GrappleStartup, CanGrapple, 1)
+            .OnEntry(_ =>
+            {
+                Animator.SetLayerWeight(1, 0);
+                _inputBuffer.ConsumeBuffer("Attack");
+                OnPlayerImpaleStateEntered?.Invoke();
+                Animator.SetTrigger("Impale");
+                _stateEntryMomentum = _momentum;
+            }).OnExit(_ => { Animator.ResetTrigger("Impale"); });
+    }
 }
