+++ Assets/Code/Fsm/GravityFsm/PlayerFsm/PlayerFsm_Jumpsquat.cs
@@ -1,0 +1,17 @@
+public partial class PlayerFsm
+{
+    private void JumpsquatConfigure()
+    {
+        Machine.Configure(PlayerFsmState.Jumpsquat)
+            .SubstateOf(GravityFsmState.Grounded)
+            .SubstateOf(PlayerFsmState.LockMomentum)
+            .PermitIf(PlayerFsmTrigger.FaceLedge, PlayerFsmState.Vault, _ => true)
+            .Permit(FsmTrigger.Timeout, PlayerFsmState.Jump)
+            .OnEntry(_ =>
+            {
+                ReplaceAnimatorTrigger("Jumpsquat");
+                _inputBuffer.ConsumeBuffer("Jump");
+            })
+            .OnExitFrom(FsmTrigger.Timeout, _ => { YVelocity = JumpYVelocity; });
+    }
+}
