+++ Assets/Code/Fsm/GravityFsm/PlayerFsm/PlayerFsm_MediumVaultHang.cs
@@ -1,0 +1,9 @@
+public partial class PlayerFsm
+{
+    private void MediumVaultHangConfigure()
+    {
+        Machine.Configure(PlayerFsmState.MediumVaultHang)
+            .SubstateOf(PlayerFsmState.VaultHang)
+            .OnEntry(_ => { ReplaceAnimatorTrigger("MediumVaultHang"); });
+    }
+}
