+++ Assets/Code/Fsm/GravityFsm/PlayerFsm/PlayerFsm_SlowVaultFinish.cs
@@ -8,4 +8,20 @@ public partial class PlayerFsm
         MoveYOntoLedge(0f, SlowVaultFinishLedgeLerpStrength);
         transform.position += transform.forward * (SlowVaultFinishForwardSpeed * Time.deltaTime);
     }
+
+    private void SlowVaultFinishConfigure()
+    {
+        Machine.Configure(PlayerFsmState.SlowVaultFinish)
+            .SubstateOf(PlayerFsmState.ForceWallRotation)
+            // .SubstateOf(PlayerFsmState.Grounded)
+            .SubstateOf(GravityFsmState.DontApplyYVelocity)
+            .Permit(FsmTrigger.Timeout, PlayerFsmState.GroundMove)
+            .Permit(GravityFsmTrigger.StartFrameGrounded, PlayerFsmState.GroundMove)
+            .OnEntry(_ =>
+            {
+                ReplaceAnimatorTrigger("SlowVaultFinish");
+                YVelocity = 0;
+            })
+            .OnExit(_ => { _momentum = 3f; });
+    }
 }
