+++ Assets/Code/Fsm/GravityFsm/PlayerFsm/PlayerFsm_SlowVaultHang.cs
@@ -1,0 +1,9 @@
+public partial class PlayerFsm
+{
+    private void SlowVaultHangConfigure()
+    {
+        Machine.Configure(PlayerFsmState.SlowVaultHang)
+            .SubstateOf(PlayerFsmState.VaultHang)
+            .OnEntry(_ => { ReplaceAnimatorTrigger("SlowVaultHang"); });
+    }
+}
