+++ Assets/Code/Fsm/GravityFsm/PlayerFsm/PlayerFsm_SlowVaultHang.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: 94f0ef70006f419bbe1a30dd7b319a7f
+timeCreated: 1759377214
