+++ Assets/Code/Fsm/GravityFsm/PlayerFsm/PlayerFsm_VaultHang.cs
@@ -6,4 +6,17 @@ public partial class PlayerFsm
         MoveYOntoLedge(VaultHangLedgeYOffset, VaultHangLedgeLerpStrength);
         HandleCollisionMove();
     }
+
+    private void VaultHangConfigure()
+    {
+        Machine.Configure(PlayerFsmState.VaultHang)
+            .SubstateOf(PlayerFsmState.ForceWallRotation)
+            .SubstateOf(GravityFsmState.DontApplyYVelocity)
+            .Permit(FsmTrigger.Timeout, PlayerFsmState.SlowVaultFinish)
+            .OnEntry(_ =>
+            {
+                if (!UpdateLedgePosition(FaceHighLedgeHeight)) UpdateLedgePosition(FaceLedgeHeight);
+                YVelocity = 0;
+            });
+    }
 }
