+++ Assets/Code/Fsm/GravityFsm/PlayerFsm/PlayerFsm_Wallsquat.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: 9fac9118596d456493470be0c9ad735a
+timeCreated: 1759377114
