+++ Assets/Code/Fsm/GravityFsm/PlayerFsm/PlayerFsm__setup.cs
@@ -7,442 +7,27 @@ public partial class PlayerFsm
         base.SetupMachine();
 
         GroundMoveConfigure();
-
         JumpsquatConfigure();
-
         LandsquatConfigure();
-
         HardLandConfigure();
-
         HardLandRollConfigure();
-        
         JumpConfigure();
-
         FallConfigure();
-
         AerialConfigure();
-
         VaultConfigure();
-
         VaultHangConfigure();
-
         SlowVaultHangConfigure();
-
         MediumVaultHangConfigure();
-
         SlowVaultFinishConfigure();
-
         WallstepConfigure();
-
         WallsquatConfigure();
-
         WallrunConfigure();
-
         DashsquatConfigure();
-
         GrappleConfigure();
-
-
         ImpaleGroundConfigure();
-
         ImpaleAirConfigure();
-
         GrappleStartupConfigure();
-
         GrappleFlipConfigure();
-    }
-
-    private void GrappleFlipConfigure()
-    {
-        Machine.Configure(PlayerFsmState.GrappleFlip)
-            .SubstateOf(GravityFsmState.Aerial)
-            .Permit(GravityFsmTrigger.StartFrameGrounded, PlayerFsmState.Landsquat)
-            .PermitIf(GravityFsmTrigger.StartFrameGrounded, PlayerFsmState.HardLand, _ => AirYDiff() < HardLandAirDiff,
-                1)
-            .PermitIf(GravityFsmTrigger.StartFrameGrounded, PlayerFsmState.HardLandRoll,
-                _ => AirYDiff() < HardLandAirDiff && _momentum > HardLandRollMinimumMomentum, 2)
-            .PermitIf(PlayerFsmTrigger.FaceLedge, PlayerFsmState.Vault, _ => YVelocity > VaultMinimumYVelocity, 1)
-            .PermitIf(PlayerFsmTrigger.FaceLedge, PlayerFsmState.MediumVaultHang, _ => true)
-            .PermitIf(PlayerFsmTrigger.FaceWall, PlayerFsmState.Wallsquat,
-                _ => _momentum > WallSquatMinimumMomentum && YVelocity < WallsquatMinimumYVelocity)
-            .PermitIf(PlayerFsmTrigger.FaceWallStrict, PlayerFsmState.Wallsquat,
-                _ => _momentum > WallSquatMinimumMomentum && YVelocity < WallsquatMinimumYVelocity)
-            .PermitIf(PlayerFsmTrigger.FaceHighLedge, PlayerFsmState.Wallsquat,
-                _ => _momentum > WallSquatMinimumMomentum && YVelocity < WallsquatMinimumYVelocity)
-            .PermitIf(PlayerFsmTrigger.FlankWall, PlayerFsmState.Wallrun,
-                _ => _momentum > WallRunMinimumMomentum && YVelocity < WallRunMinimumYVelocity)
-            .OnEntry(_ =>
-            {
-                _momentum = 10f;
-                ReplaceAnimatorTrigger("GrappleFlip");
-                YVelocity = 30;
-            });
-    }
-
-    private void GrappleStartupConfigure()
-    {
-        Machine.Configure(PlayerFsmState.GrappleStartup)
-            .SubstateOf(GravityFsmState.DontApplyYVelocity)
-            .Permit(FsmTrigger.Timeout, PlayerFsmState.Grapple)
-            .OnEntry(_ =>
-            {
-                ReplaceAnimatorTrigger("GrappleStartup");
-                YVelocity = 10;
-                _inputBuffer.ConsumeBuffer("Attack");
-            }).OnExit(_ => { YVelocity = 0; });
-
-        Machine.Configure(PlayerFsmState.GrappleFlipsquat)
-            .SubstateOf(GravityFsmState.DontApplyYVelocity)
-            .Permit(FsmTrigger.Timeout, PlayerFsmState.GrappleFlip)
-            .OnEntry(_ =>
-            {
-                // transform.DOShakePosition(0.5f, 0.3f);
-                ReplaceAnimatorTrigger("GrappleFlipsquat");
-                // HitstopManager.Singleton.StartHitstop(0.075f);
-            })
-            .OnExit(_ => { });
-    }
-
-    private void ImpaleAirConfigure()
-    {
-        Machine.Configure(PlayerFsmState.ImpaleAir)
-            .SubstateOf(GravityFsmState.Aerial)
-            .Permit(FsmTrigger.Timeout, PlayerFsmState.Fall)
-            .Permit(GravityFsmTrigger.StartFrameGrounded, PlayerFsmState.Landsquat)
-            .PermitIf(GravityFsmTrigger.StartFrameGrounded, PlayerFsmState.HardLand, _ => AirYDiff() < HardLandAirDiff,
-                1)
-            .PermitIf(GravityFsmTrigger.StartFrameGrounded, PlayerFsmState.HardLandRoll,
-                _ => AirYDiff() < HardLandAirDiff && _momentum > HardLandRollMinimumMomentum, 2)
-            .PermitIf(PlayerFsmTrigger.FaceLedge, PlayerFsmState.Vault, _ => YVelocity > VaultMinimumYVelocity, 1)
-            .PermitIf(PlayerFsmTrigger.FaceLedge, PlayerFsmState.MediumVaultHang, _ => true)
-            .PermitIf(PlayerFsmTrigger.FaceWall, PlayerFsmState.Wallsquat,
-                _ => _momentum > WallSquatMinimumMomentum && YVelocity < WallsquatMinimumYVelocity)
-            .PermitIf(PlayerFsmTrigger.FaceWallStrict, PlayerFsmState.Wallsquat,
-                _ => _momentum > WallSquatMinimumMomentum && YVelocity < WallsquatMinimumYVelocity)
-            .PermitIf(PlayerFsmTrigger.FaceHighLedge, PlayerFsmState.Wallsquat,
-                _ => _momentum > WallSquatMinimumMomentum && YVelocity < WallsquatMinimumYVelocity)
-            .PermitIf(PlayerFsmTrigger.Attack, PlayerFsmState.GrappleStartup, CanGrapple)
-            .OnEntry(_ =>
-            {
-                Animator.SetLayerWeight(1, 0);
-                _inputBuffer.ConsumeBuffer("Attack");
-                OnPlayerImpaleStateEntered?.Invoke();
-                Animator.SetTrigger("ImpaleJump");
-                _stateEntryMomentum = _momentum;
-            }).OnExit(_ => { Animator.ResetTrigger("ImpaleJump"); });
-    }
-
-    private void ImpaleGroundConfigure()
-    {
-        Machine.Configure(PlayerFsmState.ImpaleGround)
-            .SubstateOf(GravityFsmState.Grounded)
-            .Permit(FsmTrigger.Timeout, PlayerFsmState.GroundMove)
-            .Permit(GravityFsmTrigger.StartFrameAerial, PlayerFsmState.Fall)
-            .Permit(PlayerFsmTrigger.Jump, PlayerFsmState.Jumpsquat)
-            .PermitIf(PlayerFsmTrigger.Attack, PlayerFsmState.GrappleStartup, CanGrapple, 1)
-            .OnEntry(_ =>
-            {
-                Animator.SetLayerWeight(1, 0);
-                _inputBuffer.ConsumeBuffer("Attack");
-                OnPlayerImpaleStateEntered?.Invoke();
-                Animator.SetTrigger("Impale");
-                _stateEntryMomentum = _momentum;
-            }).OnExit(_ => { Animator.ResetTrigger("Impale"); });
-    }
-
-    private void GrappleConfigure()
-    {
-        Machine.Configure(PlayerFsmState.Grapple)
-            .SubstateOf(GravityFsmState.Aerial)
-            .Permit(FsmTrigger.Timeout, PlayerFsmState.GrappleFlipsquat)
-            // .Permit(PlayerFsmTrigger.ContactHitboxTrigger, PlayerFsmState.GrappleFlipsquat)
-            // .PermitIf(PlayerFsmTrigger.FaceLedge, PlayerFsmState.Vault, _ => true)
-            // .PermitIf(PlayerFsmTrigger.FaceWall, PlayerFsmState.Wallsquat, _ => true)
-            // .PermitIf(PlayerFsmTrigger.FaceWallStrict, PlayerFsmState.Wallsquat, _ => true)
-            .SubstateOf(GravityFsmState.DontApplyYVelocity)
-            .OnEntry(_ =>
-            {
-                YVelocity = 0;
-                _momentum = Mathf.Min(Mathf.Max(_momentum + DashEntryMomentumGain, DashEntryMinimumMomentum),
-                    MaxMomentum);
-                transform.rotation =
-                    Quaternion.LookRotation(PlayerWeaponFsm.Singleton.transform.position - transform.position,
-                        Vector3.up);
-                ReplaceAnimatorTrigger("Dash");
-            })
-            .OnExit(_ =>
-            {
-                transform.position = PlayerWeaponFsm.Singleton.transform.position - transform.forward * 0.75f;
-            });
-    }
-
-    private void DashsquatConfigure()
-    {
-        Machine.Configure(PlayerFsmState.Dashsquat)
-            .SubstateOf(GravityFsmState.Aerial)
-            .Permit(FsmTrigger.Timeout, PlayerFsmState.Grapple)
-            .SubstateOf(GravityFsmState.DontApplyYVelocity)
-            .OnEntry(_ =>
-            {
-                _inputBuffer.ConsumeBuffer("Dash");
-                ReplaceAnimatorTrigger("Dashsquat");
-            });
-    }
-
-    private void WallrunConfigure()
-    {
-        Machine.Configure(PlayerFsmState.Wallrun)
-            .SubstateOf(GravityFsmState.Aerial)
-            .Permit(GravityFsmTrigger.StartFrameGrounded, PlayerFsmState.Landsquat)
-            .Permit(PlayerFsmTrigger.Jump, PlayerFsmState.Jumpsquat)
-            .Permit(PlayerFsmTrigger.FlankOpen, PlayerFsmState.Fall)
-            .PermitIf(PlayerFsmTrigger.FaceLedge, PlayerFsmState.Vault, _ => YVelocity > VaultMinimumYVelocity, 1)
-            .PermitIf(PlayerFsmTrigger.FaceLedge, PlayerFsmState.MediumVaultHang, _ => true)
-            .PermitIf(PlayerFsmTrigger.FaceWallStrict, PlayerFsmState.Wallsquat,
-                _ => _momentum > WallSquatMinimumMomentum)
-            .PermitIf(PlayerFsmTrigger.FaceHighLedge, PlayerFsmState.Wallsquat,
-                _ => _momentum > WallSquatMinimumMomentum)
-            .OnEntry(_ =>
-            {
-                print("Yvelocity on entering wallrun: " + YVelocity);
-                _momentum = Mathf.Max(_momentum, WallRunMinimumEntryMomentum);
-                ReplaceAnimatorTrigger("Wallrun");
-            })
-            .OnExitFrom(PlayerFsmTrigger.Jump, _ =>
-            {
-                var rotationMod = _currentFlankType == FlankType.Left ? -1f : 1f;
-                var forward = Quaternion.Euler(0f, WallrunJumpAngle * rotationMod, 0f) * _currentFlankWallNormal;
-                transform.rotation = Quaternion.LookRotation(forward, Vector3.up);
-            })
-            .OnExit(_ => { _previousWallrunSide = _currentFlankType; });
-    }
-
-    private void WallsquatConfigure()
-    {
-        Machine.Configure(PlayerFsmState.Wallsquat)
-            .SubstateOf(GravityFsmState.Aerial)
-            .SubstateOf(PlayerFsmState.ForceWallRotation)
-            .SubstateOf(GravityFsmState.DontApplyYVelocity)
-            .Permit(GravityFsmTrigger.StartFrameGrounded, PlayerFsmState.Landsquat)
-            // .Permit(PlayerFsmTrigger.FaceOpen, PlayerFsmState.Fall)
-            .PermitIf(PlayerFsmTrigger.Jump, PlayerFsmState.Wallstep,
-                _ => TimeInCurrentState() > WallstepMinimumDuration, 1)
-            .Permit(FsmTrigger.Timeout, PlayerFsmState.Fall)
-            .OnEntry(_ =>
-            {
-                YVelocity = 0;
-                ReplaceAnimatorTrigger("Wallsquat");
-            })
-            .OnExitFrom(PlayerFsmTrigger.FaceOpen, _ => { _momentum = 0; });
-    }
-
-    private void WallstepConfigure()
-    {
-        Machine.Configure(PlayerFsmState.Wallstep)
-            .SubstateOf(GravityFsmState.Aerial)
-            .SubstateOf(PlayerFsmState.ForceWallRotation)
-            .Permit(GravityFsmTrigger.StartFrameGrounded, PlayerFsmState.Landsquat)
-            .Permit(GravityFsmTrigger.StartFrameWithNegativeYVelocity, PlayerFsmState.Fall)
-            .PermitIf(PlayerFsmTrigger.FaceHighLedge, PlayerFsmState.SlowVaultHang,
-                _ => YVelocity < MediumVaultHangMinimumYVelocity)
-            .PermitIf(PlayerFsmTrigger.FaceHighLedge, PlayerFsmState.MediumVaultHang,
-                _ => YVelocity > MediumVaultHangMinimumYVelocity, 1)
-            .PermitIf(PlayerFsmTrigger.FaceLedge, PlayerFsmState.MediumVaultHang,
-                _ => YVelocity > MediumVaultHangMinimumYVelocity, 1)
-            .OnEntry(_ =>
-            {
-                _inputBuffer.ConsumeBuffer("Jump");
-                ReplaceAnimatorTrigger("Wallstep");
-                YVelocity = Mathf.Lerp(WallstepMinimumYVelocityGain, WallstepMaximumYVelocityGain,
-                    ComputeMomentumWeight());
-                Animator.SetFloat("VerticalMomentum", ComputeMomentumWeight());
-                _momentum = 0;
-            });
-    }
-
-    private void SlowVaultFinishConfigure()
-    {
-        Machine.Configure(PlayerFsmState.SlowVaultFinish)
-            .SubstateOf(PlayerFsmState.ForceWallRotation)
-            // .SubstateOf(PlayerFsmState.Grounded)
-            .SubstateOf(GravityFsmState.DontApplyYVelocity)
-            .Permit(FsmTrigger.Timeout, PlayerFsmState.GroundMove)
-            .Permit(GravityFsmTrigger.StartFrameGrounded, PlayerFsmState.GroundMove)
-            .OnEntry(_ =>
-            {
-                ReplaceAnimatorTrigger("SlowVaultFinish");
-                YVelocity = 0;
-            })
-            .OnExit(_ => { _momentum = 3f; });
-    }
-
-    private void MediumVaultHangConfigure()
-    {
-        Machine.Configure(PlayerFsmState.MediumVaultHang)
-            .SubstateOf(PlayerFsmState.VaultHang)
-            .OnEntry(_ =>
-            {
-                ReplaceAnimatorTrigger("MediumVaultHang");
-            });
-    }
-
-    private void SlowVaultHangConfigure()
-    {
-        Machine.Configure(PlayerFsmState.SlowVaultHang)
-            .SubstateOf(PlayerFsmState.VaultHang)
-            .OnEntry(_ =>
-            {
-                ReplaceAnimatorTrigger("SlowVaultHang");
-            });
-    }
-
-    private void VaultHangConfigure()
-    {
-        Machine.Configure(PlayerFsmState.VaultHang)
-            .SubstateOf(PlayerFsmState.ForceWallRotation)
-            .SubstateOf(GravityFsmState.DontApplyYVelocity)
-            .Permit(FsmTrigger.Timeout, PlayerFsmState.SlowVaultFinish)
-            .OnEntry(_ =>
-            {
-                if (!UpdateLedgePosition(FaceHighLedgeHeight)) UpdateLedgePosition(FaceLedgeHeight);
-                YVelocity = 0;
-            });
-    }
-
-    private void VaultConfigure()
-    {
-        Machine.Configure(PlayerFsmState.Vault)
-            .SubstateOf(GravityFsmState.DontApplyYVelocity)
-            .Permit(FsmTrigger.Timeout, PlayerFsmState.Fall)
-            // .Permit(GravityFsmTrigger.StartFrameGrounded, PlayerFsmState.Landsquat)
-            .OnEntry(_ =>
-            {
-                UpdateLedgePosition(FaceLedgeHeight);
-                ReplaceAnimatorTrigger("Vault");
-                YVelocity = 0;
-            });
-    }
-
-    private void AerialConfigure()
-    {
-        Machine.Configure(GravityFsmState.Aerial)
-            .SubstateOf(PlayerFsmState.LockMomentum);
-    }
-
-    private void FallConfigure()
-    {
-        Machine.Configure(PlayerFsmState.Fall)
-            .SubstateOf(GravityFsmState.Aerial)
-            .Permit(GravityFsmTrigger.StartFrameGrounded, PlayerFsmState.Landsquat)
-            .PermitIf(PlayerFsmTrigger.Jump, PlayerFsmState.Jumpsquat, _ => TimeInAir <= CoyoteTime)
-            .PermitIf(GravityFsmTrigger.StartFrameGrounded, PlayerFsmState.HardLand, _ => AirYDiff() < HardLandAirDiff,
-                1)
-            .PermitIf(GravityFsmTrigger.StartFrameGrounded, PlayerFsmState.HardLandRoll,
-                _ => AirYDiff() < HardLandAirDiff && _momentum > HardLandRollMinimumMomentum, 2)
-            .Permit(PlayerFsmTrigger.Dash, PlayerFsmState.Dashsquat)
-            .OnEntry(_ => { ReplaceAnimatorTrigger("Fall"); });
-
-        Machine.Configure(PlayerFsmState.HardTurn)
-            .SubstateOf(PlayerFsmState.LockMomentum)
-            .Permit(PlayerFsmTrigger.NoMomentum, PlayerFsmState.GroundMove)
-            .Permit(GravityFsmTrigger.StartFrameAerial, PlayerFsmState.Fall)
-            .SubstateOf(GravityFsmState.Grounded)
-            .OnEntry(_ => { ReplaceAnimatorTrigger("HardTurn"); });
-    }
-
-    private void JumpConfigure()
-    {
-        Machine.Configure(PlayerFsmState.Jump)
-            .SubstateOf(GravityFsmState.Aerial)
-            .Permit(GravityFsmTrigger.StartFrameGrounded, PlayerFsmState.Landsquat)
-            .PermitIf(GravityFsmTrigger.StartFrameGrounded, PlayerFsmState.HardLand, _ => AirYDiff() < HardLandAirDiff,
-                1)
-            .PermitIf(GravityFsmTrigger.StartFrameGrounded, PlayerFsmState.HardLandRoll,
-                _ => AirYDiff() < HardLandAirDiff && _momentum > HardLandRollMinimumMomentum, 2)
-            .PermitIf(PlayerFsmTrigger.FaceLedge, PlayerFsmState.Vault, _ => YVelocity > VaultMinimumYVelocity, 1)
-            .PermitIf(PlayerFsmTrigger.FaceLedge, PlayerFsmState.MediumVaultHang, _ => true)
-            .PermitIf(PlayerFsmTrigger.FaceWall, PlayerFsmState.Wallsquat,
-                _ => _momentum > WallSquatMinimumMomentum && YVelocity < WallsquatMinimumYVelocity)
-            .PermitIf(PlayerFsmTrigger.FaceWallStrict, PlayerFsmState.Wallsquat,
-                _ => _momentum > WallSquatMinimumMomentum && YVelocity < WallsquatMinimumYVelocity)
-            .PermitIf(PlayerFsmTrigger.FaceHighLedge, PlayerFsmState.Wallsquat,
-                _ => _momentum > WallSquatMinimumMomentum && YVelocity < WallsquatMinimumYVelocity)
-            .PermitIf(PlayerFsmTrigger.FlankWall, PlayerFsmState.Wallrun,
-                _ => _momentum > WallRunMinimumMomentum && YVelocity < WallRunMinimumYVelocity)
-            .PermitIf(PlayerFsmTrigger.Attack, PlayerFsmState.ImpaleAir, CanImpale)
-            .PermitIf(PlayerFsmTrigger.Attack, PlayerFsmState.GrappleStartup, CanGrapple, 1)
-            // _momentum > WallRunMinimumMomentum && YVelocity < WallRunMinimumYVelocity
-            .Permit(PlayerFsmTrigger.Dash, PlayerFsmState.Dashsquat)
-            .OnEntry(_ => { ReplaceAnimatorTrigger("Jump"); });
-    }
-
-    private void HardLandRollConfigure()
-    {
-        Machine.Configure(PlayerFsmState.HardLandRoll)
-            .SubstateOf(GravityFsmState.Grounded)
-            .Permit(FsmTrigger.Timeout, PlayerFsmState.GroundMove)
-            .OnEntry(_ =>
-            {
-                _momentum = HardLandRollExitMomentum;
-                ReplaceAnimatorTrigger("HardLandRoll");
-            });
-    }
-
-    private void HardLandConfigure()
-    {
-        Machine.Configure(PlayerFsmState.HardLand)
-            .SubstateOf(GravityFsmState.Grounded)
-            .Permit(FsmTrigger.Timeout, PlayerFsmState.GroundMove)
-            .OnEntry(_ =>
-            {
-                _momentum = HardLandExitMomentum;
-                ReplaceAnimatorTrigger("HardLand");
-            });
-    }
-
-    private void LandsquatConfigure()
-    {
-        Machine.Configure(PlayerFsmState.Landsquat)
-            .SubstateOf(GravityFsmState.Grounded)
-            .SubstateOf(PlayerFsmState.LockMomentum)
-            .Permit(PlayerFsmTrigger.Jump, PlayerFsmState.Jumpsquat)
-            .Permit(FsmTrigger.Timeout, PlayerFsmState.GroundMove)
-            .OnEntry(_ => { ReplaceAnimatorTrigger("Landsquat"); })
-            .OnExit(_ =>
-            {
-                _movementAnimationMirror = !_movementAnimationMirror;
-                _previousWallrunSide = FlankType.None;
-                _currentFlankType = FlankType.None;
-                var flip = _movementAnimationMirror ? 0 : 1f;
-                Animator.SetFloat("Flip", flip);
-            });
-    }
-
-    private void JumpsquatConfigure()
-    {
-        Machine.Configure(PlayerFsmState.Jumpsquat)
-            .SubstateOf(GravityFsmState.Grounded)
-            .SubstateOf(PlayerFsmState.LockMomentum)
-            .PermitIf(PlayerFsmTrigger.FaceLedge, PlayerFsmState.Vault, _ => true)
-            .Permit(FsmTrigger.Timeout, PlayerFsmState.Jump)
-            .OnEntry(_ =>
-            {
-                ReplaceAnimatorTrigger("Jumpsquat");
-                _inputBuffer.ConsumeBuffer("Jump");
-            })
-            .OnExitFrom(FsmTrigger.Timeout, _ => { YVelocity = JumpYVelocity; });
-    }
-
-    private void GroundMoveConfigure()
-    {
-        Machine.Configure(PlayerFsmState.GroundMove)
-            .SubstateOf(GravityFsmState.Grounded)
-            .Permit(GravityFsmTrigger.StartFrameAerial, PlayerFsmState.Fall)
-            .Permit(PlayerFsmTrigger.Jump, PlayerFsmState.Jumpsquat)
-            .Permit(PlayerFsmTrigger.HardTurn, PlayerFsmState.HardTurn)
-            .PermitIf(PlayerFsmTrigger.Attack, PlayerFsmState.ImpaleGround, CanImpale)
-            .PermitIf(PlayerFsmTrigger.Attack, PlayerFsmState.GrappleStartup, CanGrapple, 1)
-            .OnEntry(_ => { ReplaceAnimatorTrigger("GroundMove"); });
     }
 
     public override void SetupStateMaps()
